Extract GCMMessageHandler to a singleton
