+++ WordPress/src/main/java/org/wordpress/android/push/NotificationsProcessingService.java
@@ -186,6 +186,7 @@ public class NotificationsProcessingService extends Service {
 
     private class QuickActionProcessor {
         private SystemNotificationsTracker mSystemNotificationsTracker;
+        private GCMMessageHandler mGCMMessageHandler;
         private String mNoteId;
         private String mReplyText;
         private String mActionType;
@@ -218,7 +219,7 @@ public class NotificationsProcessingService extends Service {
                             NotificationPushIds.AUTH_PUSH_NOTIFICATION_ID, mContext);
                     NativeNotificationsUtils.dismissNotification(
                             NotificationPushIds.ACTIONS_PROGRESS_NOTIFICATION_ID, mContext);
-                    GCMMessageService.removeNotification(NotificationPushIds.AUTH_PUSH_NOTIFICATION_ID);
+                    mGCMMessageHandler.removeNotification(NotificationPushIds.AUTH_PUSH_NOTIFICATION_ID);
 
                     AnalyticsTracker.track(AnalyticsTracker.Stat.PUSH_AUTHENTICATION_IGNORED);
                     return;
@@ -233,13 +234,13 @@ public class NotificationsProcessingService extends Service {
                     Log.d("stats",
                             "Dismissing notification with id: " + notificationId + " and type: " + mNotificationType);
                     if (notificationId == GROUP_NOTIFICATION_ID) {
-                        GCMMessageService.clearNotifications();
+                        mGCMMessageHandler.clearNotifications();
                     } else if (notificationId == QUICK_START_REMINDER_NOTIFICATION_ID) {
                         AnalyticsTracker.track(Stat.QUICK_START_NOTIFICATION_DISMISSED);
                     } else {
-                        GCMMessageService.removeNotification(notificationId);
+                        mGCMMessageHandler.removeNotification(notificationId);
                         // Dismiss the grouped notification if a user dismisses all notifications from a wear device
-                        if (!GCMMessageService.hasNotifications()) {
+                        if (!mGCMMessageHandler.hasNotifications()) {
                             NotificationManagerCompat notificationManager = NotificationManagerCompat.from(mContext);
                             notificationManager.cancel(GROUP_NOTIFICATION_ID);
                         }
@@ -448,7 +449,7 @@ public class NotificationsProcessingService extends Service {
                     mContext,
                     NotificationType.ACTIONS_PROGRESS);
             // remove the original notification from the system bar
-            GCMMessageService.removeNotificationWithNoteIdFromSystemBar(mContext, mNoteId);
+            mGCMMessageHandler.removeNotificationWithNoteIdFromSystemBar(mContext, mNoteId);
 
             // after 3 seconds, dismiss the notification that indicated success
             Handler handler = new Handler(getMainLooper());
@@ -650,7 +651,7 @@ public class NotificationsProcessingService extends Service {
         }
 
         private void resetOriginalNotification() {
-            GCMMessageService.rebuildAndUpdateNotificationsOnSystemBarForThisNote(mContext, mNoteId);
+            mGCMMessageHandler.rebuildAndUpdateNotificationsOnSystemBarForThisNote(mContext, mNoteId);
         }
     }
